chore(tests): catch operator logs during the tests run
Since we included the catch all for the logs of all the pods and show that it works,
now we can do the same for the operator logs using the same functions and model.

We gather the operator pods by matching the labels and we split them by different
pod name, meaning that if the operator is destroyed and recreated we will have
a different log file for the new pod.

Also, some clean up of repeated tailing logs to avoid more queries to the API and
unnecesary logs requirements.

Closes #5453

Signed-off-by: Jonathan Gonzalez V. <jonathan.gonzalez@enterprisedb.com>